Resolve upstream conflicts
